rename: 스토어 파일명 수정 (#29)
- 기존 LikeStore에서 useLikeStore로 파일명 수정
